<fix>:move AdminNoteDetailDTO to appropriate directory
